Rename energy normalization.
