+++ rdtools/__init__.py
@@ -1,4 +1,4 @@
-from energy_normalization import normalize_with_sapm
+from normalization import normalize_with_sapm
 from degradation import degradation_with_ols
 
 from ._version import get_versions
